refactor(handler): extract DoDelete from DeleteWeb
